Major change to the timeout system
Now OCaml knows nothing anymore of how long it should run, and python is
handling this. So, OCaml runs "forever" until one of the following
happens:
* It reaches the frontier
* It receives a SIGUSR1
In both cases it terminates cleanly by sending back the program found so
far - if any.
Python handles the timeout: it waits for a thread to end, and if it
doesn't it sends the SIGUSR1 and gather the result to return it.

This avoid all the fiddling around with the "Time" module in OCaml. As
far I can tell this works with the grom project, and does not break the
list one.